Minor cleanup and code reorganization
